+++ CHANGELOG.md
@@ -21,8 +21,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `OxidEsales\TestingLibrary\ModuleLoader::useOriginalChain()`
 
 ## [v7.3.0] - Unreleased
+
 ### Added 
 - Support of codeception v4
+
+### Deprecated
+- bin/runtests_coverage
+- \OxidEsales\TestingLibrary\Services\Library\ProjectConfigurationHandler::cleanup()
+- \OxidEsales\TestingLibrary\library\Services\ProjectConfiguration\ProjectConfiguration
 
 ## [v7.2.0] - 2020-11-10